+++ spec/spec_helper.rb
@@ -6,6 +6,10 @@ require 'puppetlabs_spec_helper/module_spec_helper'
 require 'rspec-puppet-facts'
 include RspecPuppetFacts
 
+                                                    # Original fact sources:
+add_custom_fact :concat_basedir, '/tmp'             # puppetlabs-concat
+add_custom_fact :puppetversion, Puppet.version      # Facter, but excluded from rspec-puppet-facts
+
 # Workaround for no method in rspec-puppet to pass undef through :params
 class Undef
   def inspect; 'undef'; end
@@ -29,44 +33,32 @@ def exclude_test_os
   end
 end
 
+# Use the above environment variables to limit the platforms under test
+def on_os_under_test
+  on_supported_os.reject do |os, facts|
+    (only_test_os() && !only_test_os.include?(os)) ||
+      (exclude_test_os() && exclude_test_os.include?(os))
+  end
+end
+
 def get_content(subject, title)
+  is_expected.to contain_file(title)
   content = subject.resource('file', title).send(:parameters)[:content]
   content.split(/\n/).reject { |line| line =~ /(^#|^$|^\s+#)/ }
 end
 
 def verify_exact_contents(subject, title, expected_lines)
-  expect(get_content(subject, title)).to eq(expected_lines)
+  expect(get_content(subject, title)).to match_array(expected_lines)
 end
 
 def verify_concat_fragment_contents(subject, title, expected_lines)
+  is_expected.to contain_concat__fragment(title)
   content = subject.resource('concat::fragment', title).send(:parameters)[:content]
-  expect(content.split("\n") & expected_lines).to eq(expected_lines)
+  expect(content.split("\n") & expected_lines).to match_array(expected_lines)
 end
 
 def verify_concat_fragment_exact_contents(subject, title, expected_lines)
+  is_expected.to contain_concat__fragment(title)
   content = subject.resource('concat::fragment', title).send(:parameters)[:content]
-    expect(content.split(/\n/).reject { |line| line =~ /(^#|^$|^\s+#)/ }).to eq(expected_lines)
-end
-
-# See https://github.com/rodjek/rspec-puppet/issues/329
-# Without this patch the @@cache variable grows huge and causes memory usage issues
-# with a large amount of examples.
-module RSpec::Puppet
-  module Support
-    def build_catalog(*args)
-      if @@cache.has_key?(args)
-        @@cache[args]
-      else
-        @@cache = {}
-        @@cache[args] ||= self.build_catalog_without_cache(*args)
-      end
-    end
-  end
-end
-
-# See https://github.com/rodjek/rspec-puppet/pull/333
-# Prevents each generated catalog being held in memory after the example group has
-# completed.
-RSpec.configure do |c|
-  c.after(:each) { @catalogue = nil }
+  expect(content.split(/\n/).reject { |line| line =~ /(^#|^$|^\s+#)/ }).to match_array(expected_lines)
 end